Decompose file into separate modules
To ease future refactors on the code, decompose the file into a module
with multiple modules that can be refactored semi-independently.
