+++ wqflask/base/data_set/tempdataset.py
@@ -1,0 +1,23 @@
+"TempDataSet class ..."
+
+from .dataset import DataSet
+
+class TempDataSet(DataSet):
+    """Temporary user-generated data set"""
+
+    def setup(self):
+        self.search_fields = ['name',
+                              'description']
+
+        self.display_fields = ['name',
+                               'description']
+
+        self.header_fields = ['Name',
+                              'Description']
+
+        self.type = 'Temp'
+
+        # Need to double check later how these are used
+        self.id = 1
+        self.fullname = 'Temporary Storage'
+        self.shortname = 'Temp'
